# .NET 8.0 and a few other perf improvements