… .github/workflows/ci.yml
… src/Markdig/Parsers/Inlines/CodeInlineParser.cs
… src/Markdig/Polyfills/EncodingExtensions.cs
… src/Markdig/Polyfills/IndexOfHelpers.cs
… src/Markdig/Polyfills/SearchValues.cs
… src/Markdig/Renderers/HtmlRenderer.cs
… src/Markdig/Syntax/Inlines/ContainerInline.cs
… src/Markdig/Syntax/Inlines/EmphasisDelimiterInline.cs
… src/Markdig/Syntax/Inlines/Inline.cs
… src/Markdig/Syntax/MarkdownObject.cs
… src/global.json